Reskins the entire dashboard to a light-on-dark theme
The previous patch "inverted" the geo graph theme to
make the colors (which were also changed by the patch
to polloi) pop out better. This resulted in the new
geo graph looking really out of place, and it clashed
with the white bar at the top of the dashboard. Thus,
this patch reskins the rest of the dashboard to that
theme, making the geo graph appear in unison with the
rest while also differentiating the Portal dashboard
visually from the other Discovery dashboard, which is
appropriate given the uniqueness of the project/team.

Bug: T123959
Change-Id: Id0ed69719b04f1d0590ed0e22ff459ef253a3638